added new enemy assets, implemented bullet,

+++ Assets/Scripts/ScriptableObjects/EnemyStats.cs
@@ -7,23 +7,23 @@ namespace ScriptableObjects
     public class EnemyStats : ScriptableObject
     {
         [SerializeField] private int health;
-        [SerializeField] private int speed;
-        [SerializeField] private int attackDamage;
-        [SerializeField] private int attackRange;
-        [SerializeField] private int attackSpeed;
-        [SerializeField] private int chaseRange;
-        [SerializeField] private int deathTime;
-        [SerializeField] private int attackCooldown;
+        [SerializeField] private float speed;
+        [SerializeField] private float attackDamage;
+        [SerializeField] private float attackRange;
+        [SerializeField] private float attackSpeed;
+        [SerializeField] private float chaseRange;
+        [SerializeField] private float deathTime;
+        [SerializeField] private float attackCooldown;
         [SerializeField] private GameObject bulletPrefab;
         
         public int Health => health;
-        public int Speed => speed;
-        public int AttackDamage => attackDamage;
-        public int AttackRange => attackRange;
-        public int AttackSpeed => attackSpeed;
-        public int ChaseRange => chaseRange;
-        public int DeathTime => deathTime;
-        public int AttackCooldown => attackCooldown;
+        public float Speed => speed;
+        public float AttackDamage => attackDamage;
+        public float AttackRange => attackRange;
+        public float AttackSpeed => attackSpeed;
+        public float ChaseRange => chaseRange;
+        public float DeathTime => deathTime;
+        public float AttackCooldown => attackCooldown;
         public GameObject BulletPrefab => bulletPrefab;
         
         public UnityAction<float> OnHealthChanged;
